+++ packages/grafana-runtime/src/utils/queryResponse.ts
@@ -14,6 +14,7 @@ import {
   dataFrameFromJSON,
 } from '@grafana/data';
 import { FetchError, FetchResponse } from '../services';
+import { toDataQueryError } from './toDataQueryError';
 
 /**
  * Single response object from a backend data source. Properties are optional but response should contain at least
@@ -157,36 +158,6 @@ export function toTestingStatus(err: FetchError): any {
   }
 
   throw err;
-}
-
-/**
- * Convert an object into a DataQueryError -- if this is an HTTP response,
- * it will put the correct values in the error field
- *
- * @public
- */
-export function toDataQueryError(err: DataQueryError | string | Object): DataQueryError {
-  const error = (err || {}) as DataQueryError;
-
-  if (!error.message) {
-    if (typeof err === 'string' || err instanceof String) {
-      return { message: err } as DataQueryError;
-    }
-
-    let message = 'Query error';
-    if (error.message) {
-      message = error.message;
-    } else if (error.data && error.data.message) {
-      message = error.data.message;
-    } else if (error.data && error.data.error) {
-      message = error.data.error;
-    } else if (error.status) {
-      message = `Query error: ${error.status} ${error.statusText}`;
-    }
-    error.message = message;
-  }
-
-  return error;
 }
 
 /**
