+++ packages/grafana-runtime/src/utils/toDataQueryError.ts
